another minor fix

+++ utils/get_torch_data_loader.py
@@ -28,7 +28,7 @@ class SustainBenchTextTorchDataset(Dataset):
         dataset = SustainBenchTextDataset(features=features, target=target, model_type=model_type.name, classification_threshold=classification_threshold, data_dir=data_dir, split_scheme=split_scheme)
         self.embeddings,self.labels = dataset.get_data(data_split)
         self.embeddings=torch.from_numpy(self.embeddings).float()
-        self.labels = torch.from_numpy(self.labels).long()
+        self.labels = torch.from_numpy(self.labels)
 
 
     def __len__(self):
